refactor(context): update GlobalContext import paths for consistency
